Fix docs build
Add missing imports.
Move exceptions to the central place.
